Tests for concrete classes that have dependencies injected via the constructor.
When Castle creates a ClassProxy, it tries to invoke a ctor for the proxy with a param list that matches that of the concrete class ctor.  If you choose not to pass any ctor args via the proxy, Castle will not be able to find a matching ctor to call.

Since the args are not actually used by proxy code, it is safe to simply pass null for each argument.

Struct params are therefore not supported (yet).